fix and docs: refactor codes and update documentation
