+++ src/js/displayitems.js
@@ -3,7 +3,7 @@ import addItemToUi from './additemtoui';
 
 const displayItems = databaseName => {
   const items = getItems(databaseName);
-  items.forEach(item => addItemToUi(item, databaseName, 'add'));
+  items.forEach(item => addItemToUi(item)[databaseName].add());
 };
 
 export default displayItems;
