+++ src/js/filter.js
@@ -3,10 +3,13 @@ import addItemToUi from './additemtoui';
 /* eslint no-unused-expressions: ["error", { "allowTernary": true }] */
 
 const listToShow = items => {
+  const todoRow = document.querySelector('#todoRow');
+  todoRow.querySelectorAll('*').forEach(row => row.remove());
   items.forEach(item => {
-    addItemToUi(item, 'todo', 'add');
+    addItemToUi(item).todo.add();
   });
 };
+
 const filterTodo = ({
   target: {
     classList,
@@ -15,10 +18,7 @@ const filterTodo = ({
 }) => {
   if (classList.contains('filter')) {
     const todos = getItems('todo');
-
     const filteredArrays = todos.filter(({ selectProject }) => selectProject === tofilter);
-    const todoRow = document.querySelector('#todoRow');
-    todoRow.querySelectorAll('*').forEach(row => row.remove());
     tofilter === 'all' ? listToShow(todos) : listToShow(filteredArrays);
   }
 };
