+++ src/js/getitems.js
@@ -1,10 +1,6 @@
+let items;
 const getItems = databaseName => {
-  let items;
-  /* eslint no-unused-expressions: ["error", { "allowTernary": true }] */
-  localStorage.getItem(databaseName) === null
-    ? (items = [])
-    : (items = JSON.parse(localStorage.getItem(databaseName)));
-
+  items = localStorage.getItem(databaseName) ? JSON.parse(localStorage.getItem(databaseName)) : [];
   return items;
 };
