+++ src/views/index.html
@@ -150,7 +150,7 @@
                         <span aria-hidden="true">&times;</span>
                       </button>
                     </div>
-                    <form id="projectForm" data-databasename="project">
+                    <form id="projectForm" data-databasename="project" data-formtype="add">
                       <div class="modal-body">
                         <div class="alert d-none alert-danger" role="alert" id="alert"></div>
                         <div class="form-group">
